Put generated StaConfig.hh into build artifacts.
Before, the cmake configure_file() for the StaConfig.hh header, that
make only sense for the particular cmake configuration as it contains
specific defines for that build configuration, ended up in the
original source tree.

With this fix, it is a build artifact inside build/.

Tested, that the final `make install` properly copies the
StaConfig.hh to the correct location.

Signed-off-by: Henner Zeller <hzeller@google.com>