sdc/liberty: more properties
* cells:
  * design_type: cell, macro, or module
  * is_physical_only: crudely approximated by checking for a lack of signal IOs
  * is_clock_pin
  * is_rise_edge_triggered, is_fall_edge_triggered: for outputs

* lib cells:
  * design_type: cell, or macro
  * has_timing_model: checks if the liberty cell has any timing arcs
  * is_integrated_clock_gating_cell
  * is_sequential: checks if cell has any sequentials
  * dont_use
  * valid_purposes: empty string (for now)

also, implemented the following functions:

* all_inputs -exclude_clock_ports (option is alias for -no_clocks)
* get_object_name (alias of get_full_name)
* set_dont_use: mutates LibertyCell's dont_use property

  ---

Additionally, I've added these properties as methods for the Tcl object
as well, which OpenSTA in general is quite inconsistent about. Ideally
there'd be some kind of dynamic method resolver that would resolve
properties as methods, but that would require SWIG to be more flexible
than it is.