fix: deduplication affecting sorts
This commit ensures that post-deduplication, the paths are returned in the same exact sort as previously.

The complication was the `last` boolean introduced for JSON reports - it would be O(n²) to know ahead of time if you're the last endpoint as you would have to make sure every subsequent path is qualified to be printed, i.e., not a duplicate.

This is avoided in this commit by prefixing a comma to all paths except the first path instead of suffix commas to all paths except the last path — being the first path can be trivially detected by the already existing variable `prev_end` pointer being null.

This in turn allows for a simpler deduplication algorithm: a primary iteration is used to build a set of "qualified" endpoints within a Set, and in the second iteration, if an endpoint is qualified, it's printed and `prev_end` is updated. This is O(n log n) as accesses to Set are O(log n).